Update README with GitHub notation
Because this gem isn't published on RubyGem.org (yet) the Gemfile needs
to let Bundler know where to find the Gem. Also, because we are setting
this up as a standardization tool, and we use Bundler exclusively, there
is no need for instructions on how to install the Gem manaually --
especially since those directions won't work.